fix: use timestamp_ms mode and drop Postgres cast syntax
- db/schema.ts: integer({ mode: "timestamp" }) → timestamp_ms so seed.sqlite millisecond values parse correctly
- app/(app)/projects/page.tsx: remove ::int Postgres cast from sql`count()` expressions (SQLite doesn't support it)

+++ app/(app)/projects/page.tsx
@@ -20,9 +20,9 @@ export default async function ProjectsPage({ searchParams }: { searchParams: Sea
         createdAt: projects.createdAt,
         updatedAt: projects.updatedAt,
         archivedAt: projects.archivedAt,
-        issueCount: sql<number>`count(${issues.id})::int`,
-        openIssueCount: sql<number>`count(case when ${issues.archivedAt} is null and ${issues.status} != 'done' then 1 end)::int`,
-        doneActiveCount: sql<number>`count(case when ${issues.archivedAt} is null and ${issues.status} = 'done' then 1 end)::int`,
+        issueCount: sql<number>`count(${issues.id})`,
+        openIssueCount: sql<number>`count(case when ${issues.archivedAt} is null and ${issues.status} != 'done' then 1 end)`,
+        doneActiveCount: sql<number>`count(case when ${issues.archivedAt} is null and ${issues.status} = 'done' then 1 end)`,
       })
       .from(projects)
       .leftJoin(issues, eq(issues.projectId, projects.id))
